Added static token
